+++ docs/changelog.rst
@@ -21,6 +21,9 @@ To avail of fixes in an unreleased version, please download a ZIP file
 In progress (unreleased)
 ------------------------
 
+* :gh:issue:`1451` :mod:`mitogen`: Refactor module whitelist & blacklist with
+  module overrides and blocks. Improve error messages for denied modules.
+
 
 v0.3.41 (2026-02-10)
 --------------------